+++ pupa/scrape/schemas/vote_event.py
@@ -15,6 +15,7 @@ schema = {
         'organization': {"type": ["string", "null"]},
         'legislative_session': {"type": "string"},
         'bill': {"type": ["string", "null"]},
+        'bill_action': {"type": ["string", "null"]},
         'votes': {
             "items": {
                 "type": "object",